refactor: Fix a typo: collecter -> collector
