+++ src/main/kotlin/com/expediagroup/sdk/core/configuration/collector/ConfigurationCollector.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.expediagroup.sdk.core.configuration.collecter
+package com.expediagroup.sdk.core.configuration.collector
 
 import com.expediagroup.sdk.core.configuration.provider.ConfigurationProvider
